feat: add AppLayout
