+++ src/App.jsx
@@ -1,21 +1,22 @@
 import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import AppLayout from 'components/AppLayout/AppLayout.jsx';
 import { CartProvider } from 'contexts/CartContext.jsx';
+import { AuthProvider } from 'contexts/AuthContext.jsx';
 import ProtectedRoute from 'router/ProtectedRoute.jsx';
 import routes from 'router/routerConfig.js';
-import { AuthProvider } from 'contexts/AuthContext.jsx';
 
 const App = () => (
     <AuthProvider>
         <CartProvider>
             <Router>
-                <div className="App">
+                <AppLayout>
                     <Routes>
                         {routes.map(({ path, element, protected: protectionType }) => (
                             <Route key={path} path={path} element={<ProtectedRoute element={element} protectionType={protectionType} />} />
                         ))}
                     </Routes>
-                </div>
+                </AppLayout>
             </Router>
         </CartProvider>
     </AuthProvider>
