+++ src/pages/CartPage.jsx
@@ -1,12 +1,3 @@
-import AppHeader from 'components/AppHeader/AppHeader.jsx';
-import AppFooter from 'components/AppFooter/AppFooter.jsx';
-
-const CartPage = () => (
-    <>
-        <AppHeader />
-        There will be Cart Page
-        <AppFooter />
-    </>
-);
+const CartPage = () => <>There will be Cart Page</>;
 
 export default CartPage;
